[REF] mail, im_livechat: move author open chat to discuss folder
Part of task-3265211
