+++ addons/mail/static/src/core/common/message.xml
@@ -15,11 +15,8 @@
                 <MessageInReply t-if="message.parentMessage" alignedRight="isAlignedRight" message="message" onClick="props.onParentMessageClick"/>
                 <div class="o-mail-Message-core position-relative d-flex flex-shrink-0">
                     <div class="o-mail-Message-sidebar d-flex flex-shrink-0" t-att-class="{ 'justify-content-end': isAlignedRight, 'align-items-center': props.squashed }">
-                        <div t-if="!props.squashed" class="o-mail-Message-avatarContainer position-relative bg-view" t-on-click="ev => this.onClickAuthor(ev)">
-                            <img class="o-mail-Message-avatar w-100 h-100 rounded o_object_fit_cover"
-                                t-att-title="authorText"
-                                t-att-src="authorAvatarUrl"
-                                t-att-class="{ 'o_redirect cursor-pointer': hasAuthorClickable }"/>
+                        <div t-if="!props.squashed" class="o-mail-Message-avatarContainer position-relative bg-view">
+                            <img class="o-mail-Message-avatar w-100 h-100 rounded o_object_fit_cover" t-att-src="authorAvatarUrl"/>
                         </div>
                         <t t-elif="!message.isTransient">
                             <small t-if="isActive" class="o-mail-Message-date text-muted opacity-50 ms-2">
@@ -35,7 +32,7 @@
                     </div>
                     <div class="w-100 o-min-width-0" t-att-class="{ 'flex-grow-1': message.composer }" t-ref="messageContent">
                         <div t-if="!props.squashed" class="o-mail-Message-header d-flex flex-wrap align-items-baseline mb-1 lh-1">
-                            <span t-if="(message.author or message.email_from) and shouldDisplayAuthorName" class="o-mail-Message-author" t-att-class="{ 'cursor-pointer o_redirect': hasAuthorClickable }" t-att-title="authorText" t-on-click="ev => this.onClickAuthor(ev)">
+                            <span t-if="(message.author or message.email_from) and shouldDisplayAuthorName" class="o-mail-Message-author">
                                 <strong class="me-1 text-truncate"><t t-if="message.author" t-esc="message.author.name"/><t t-else="" t-esc="message.email_from"/></strong>
                             </span>
                             <small t-if="!message.isTransient" class="o-mail-Message-date text-muted opacity-50" t-att-class="{ 'me-2': !isAlignedRight }" t-att-title="message.datetimeShort">
